Divide algorithms into separate modules
