+++ src/Makefile
@@ -1,7 +1,8 @@
 # CXX=g++
 CXX=clang++
 
-OBJ=main.o straceparser.o ids.o argparse.o states.o optimizer.o algorithms.o
+OBJ=main.o straceparser.o ids.o argparse.o states.o optimizer.o algorithm.o\
+	algorithm_weak.o algorithm_strict.o algorithm_advanced.o
 BIN=st2se
 LDFLAGS=-Iinclude/
 LINK.o = $(LINK.cpp)
@@ -66,7 +67,9 @@ clean:
 	-rm -f *.gcda *.gcno coverage.info
 
 # DEPS: generated with 'dep' recipe
-algorithms.o: algorithms.cpp algorithms.hpp ids.hpp
+algorithm.o: algorithm.cpp algorithm.hpp ids.hpp
+algorithm_weak.o: algorithm_weak.cpp algorithm_weak.hpp ids.hpp \
+ algorithm.hpp
 argparse.o: argparse.cpp argparse.hpp
 ids.o: ids.cpp ids.hpp
 main.o: main.cpp include/tao/pegtl.hpp include/tao/pegtl/config.hpp \
@@ -195,7 +198,7 @@ main.o: main.cpp include/tao/pegtl.hpp include/tao/pegtl/config.hpp \
  include/tao/pegtl/string_input.hpp \
  include/tao/pegtl/internal/file_reader.hpp \
  include/tao/pegtl/string_input.hpp argparse.hpp ids.hpp optimizer.hpp \
- algorithms.hpp states.hpp straceparser.hpp \
+ algorithm.hpp states.hpp straceparser.hpp \
  include/tao/pegtl/contrib/tracer.hpp \
  include/tao/pegtl/contrib/../config.hpp \
  include/tao/pegtl/contrib/../normal.hpp \
@@ -205,8 +208,9 @@ main.o: main.cpp include/tao/pegtl.hpp include/tao/pegtl/config.hpp \
  include/tao/pegtl/analysis/../config.hpp \
  include/tao/pegtl/analysis/grammar_info.hpp \
  include/tao/pegtl/analysis/insert_guard.hpp straceparser_actions.hpp \
- straceparser_grammar.hpp
-optimizer.o: optimizer.cpp optimizer.hpp algorithms.hpp ids.hpp
+ straceparser_grammar.hpp algorithm_weak.hpp algorithm_strict.hpp \
+ algorithm_advanced.hpp
+optimizer.o: optimizer.cpp optimizer.hpp algorithm.hpp ids.hpp
 states.o: states.cpp states.hpp ids.hpp
 straceparser.o: straceparser.cpp straceparser.hpp include/tao/pegtl.hpp \
  include/tao/pegtl/config.hpp include/tao/pegtl/version.hpp \
