+++ src/algorithm_strict.cpp
@@ -1,0 +1,12 @@
+#include "algorithm_strict.hpp"
+
+namespace st2se {
+
+	bool Algo_strict::optimize(Ids &in, Ids &out) {
+        (void) in;
+        (void) out;
+        std::cout << "Algo_strict optimize emitted." << std::endl;
+        return true;
+    }
+
+} // namespace st2se
